rename common.js
