+++ core-ui/src/routing/clusterResources/customResourceDefinitionsCl.routes.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { createResourceRoutes } from '../common';
+import { createResourceRoutes } from 'routing/createResourceRoutes';
 
 const List = React.lazy(() =>
   import('../../components/Predefined/List/CustomResourceDefinitions.list'),
