+++ core-ui/src/routing/createResourceRoutes.js
