+++ core-ui/src/routing/other.js
@@ -1,14 +1,14 @@
 import React from 'react';
 
-import noPermissions from './other/noPermissions';
-import nodeOverview from './other/nodeOverview';
-import clusterList from './other/clusters';
-import preferences from './other/preferences';
-import appServiceDetails from './other/appServiceDetails';
-import containerDetails from './other/containerDetails';
-import helmReleasesList from './other/helmReleasesList';
-import helmReleaseDetails from './other/helmReleaseDetails';
-import CRDetails from './other/CRDetails';
+import noPermissions from './other/noPermissions.routes';
+import nodeOverview from './other/nodeOverview.routes';
+import clusterList from './other/clusters.routes';
+import preferences from './other/preferences.routes';
+import appServiceDetails from './other/appServiceDetails.routes';
+import containerDetails from './other/containerDetails.routes';
+import helmReleasesList from './other/helmReleasesList.routes';
+import helmReleaseDetails from './other/helmReleaseDetails.routes';
+import CRDetails from './other/CRDetails.routes';
 
 const other = (
   <>
